[Feature]: Added LINQ and List examples
Added examples for Ex0 and Ex1. Employee class is utilized to create employees and perform filtering based on specific criteria.

Added example for Ex2. Filtering is performed on an array using the "where" clause.

Added examples for Ex3, Ex4, and Ex5. Sorting and filtering of array elements are demonstrated.

Added examples for Ex6 and Ex7. Various operations are performed on a List<string>.